cafe-recommend/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
python-dotenv==1.0.0
alembic==1.12.1
langchain==0.0.350
chromadb==0.4.22
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
python-magic==0.4.27 